Flatten local-live-view module structure, remove wrapper classes
