+++ local-live-view/assets/local_live_view/index.ts
@@ -1,21 +1,95 @@
 import { Popcorn } from "@swmansion/popcorn";
 import type { Channel } from "phoenix";
 import type { Hook, LiveSocketInstanceInterface } from "phoenix_live_view";
-import type { LLVConfig, LLVSocket, LLVView, LLVServerMessageDetail } from "./types";
+import type { LLVConfig, LLVSocket, LLVView, LLVServerMessageDetail, ViewRegistry } from "./types";
 import { setupFakeView } from "./view_setup";
 import { registerNavigationHandlers } from "./navigation";
 import { registerCustomEventBindings } from "./events";
 import { withHostLV, sendServerMessage, parseAssigns, resolveLlvId } from "./helpers";
-import { PopcornClient } from "./core/popcorn_client";
-import { ViewRegistry } from "./core/view_registry";
 
 export type { LLVConfig };
+export type { PopcornClient };
+
+const DEFAULT_CALL_TIMEOUT_MS = 10_000;
+
+type CallResult = Awaited<ReturnType<Popcorn["call"]>>;
+
+interface CreateArgs {
+  id: string;
+  view: string | null;
+  url: string;
+  urlParams: Record<string, string>;
+  assigns: Record<string, unknown>;
+}
+
+class PopcornClient {
+  private popcorn: Popcorn | null = null;
+
+  get ready(): boolean {
+    return this.popcorn !== null;
+  }
+
+  attach(popcorn: Popcorn): void {
+    this.popcorn = popcorn;
+  }
+
+  private call(args: Record<string, unknown>): Promise<CallResult> {
+    if (!this.popcorn) {
+      return Promise.reject(new Error("LLV: PopcornClient used before runtime was ready"));
+    }
+    return this.popcorn.call(args, { timeoutMs: DEFAULT_CALL_TIMEOUT_MS });
+  }
+
+  private fire(action: string, args: Record<string, unknown>): void {
+    this.call(args).then(
+      (result) => {
+        if (!result.ok) console.error(`LLV ${action} error`, result.error);
+      },
+      (err) => console.error(`LLV ${action} error`, err),
+    );
+  }
+
+  create({ id, view, url, urlParams, assigns }: CreateArgs): Promise<CallResult> {
+    return this.call({
+      action: "create",
+      id,
+      view,
+      url,
+      url_params: urlParams,
+      assigns,
+    });
+  }
+
+  destroy(id: string): void {
+    this.fire("destroy", { action: "destroy", id, payload: {} });
+  }
+
+  reconnected(id: string): void {
+    this.fire("reconnect sync", { action: "reconnected", id, payload: {} });
+  }
+
+  updateAssigns(id: string, assigns: Record<string, unknown>): void {
+    this.fire("update assigns", { action: "update_assigns", id, assigns });
+  }
+
+  handleParams(id: string, params: Record<string, string>, url: string): void {
+    this.fire("handle_params", { action: "handle_params", id, payload: { params, url } });
+  }
+
+  event(id: string, payload: Record<string, unknown>): void {
+    this.fire("event", { action: "event", id, payload });
+  }
+
+  push(id: string, event: string, payload: Record<string, unknown>): Promise<CallResult> {
+    return this.call({ action: "push", id, payload: { event, payload } });
+  }
+}
 
 export class LLVEngine {
   private socket: LLVSocket;
   private config: LLVConfig;
   private pop = new PopcornClient();
-  private views = new ViewRegistry();
+  private views: ViewRegistry = new Map();
   private channels: Record<string, Channel> = {};
   private bufferedServerMessages: LLVServerMessageDetail[] = [];
 
@@ -88,8 +162,9 @@ export class LLVEngine {
   // the host LiveView removes the mount point.
   private unmountView(pop_view_el: HTMLElement): void {
     const llvId = pop_view_el.id;
-    const view = this.views.remove(llvId);
+    const view = this.views.get(llvId);
     if (!view) return;
+    this.views.delete(llvId);
     view.destroy?.();
     this.pop.destroy(llvId);
   }
@@ -208,7 +283,7 @@ export class LLVEngine {
     window.__popcornTransportReceive = (llvId: string, diff: Record<string, unknown>) => {
       const view = this.views.get(llvId);
       if (!view) {
-        console.error("LLV view not found:", llvId, "available:", this.views.ids());
+        console.error("LLV view not found:", llvId, "available:", this.views.keys());
         return;
       }
       view.update(diff, []);
